+++ src/components/forms/FormFieldsEditor.tsx
@@ -1,44 +1,26 @@
 import { useEffect, useState, useRef } from 'react';
 import { Form, TextFormField, SelectFormField } from 'types';
-import { FormFieldType, FormResourceType } from 'types/Enums';
-import { useUpdateForm, useCreateForm, useFormSubmissions } from 'hooks/Form';
-import { FormCreate, SelectFormFieldOption } from 'types/Form';
-import { useSnackbar } from 'hooks/Snackbar';
-import { ClickAwayListener, Grow, Paper, Popper, Typography, MenuItem, MenuList, Button, Stack, TextField } from '@mui/material';
+import { FormFieldType } from 'types/Enums';
+import { useUpdateForm, useFormSubmissions } from 'hooks/Form';
+import { ClickAwayListener, Grow, Paper, Popper, Typography, MenuItem, MenuList, Button, Stack } from '@mui/material';
 import FieldEditor from 'components/forms/FieldEditor';
-import VerifyDialog from 'components/layout/VerifyDialog';
+import { useSnackbar } from 'hooks/Snackbar';
 
 export type FormFieldsEditorProps = {
   form: Form;
+  onSave?: () => void;
+  newTitle?: string;
 };
 
-const removeIdsFromFields = (fields: Array<TextFormField | SelectFormField>) => {
-  const newFields: Array<TextFormField | SelectFormField> = [];
-  fields.forEach((field) => {
-    const { id, ...restField } = field;
-    const newOptions: Array<SelectFormFieldOption> = [];
-
-    field.options.forEach((option) => {
-      const { id, ...restOption } = option;
-      newOptions.push(restOption as SelectFormFieldOption);
-    });
-
-    newFields.push({ ...restField, options: newOptions } as TextFormField | SelectFormField);
-  });
-  return newFields;
-};
-
-const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
+const FormFieldsEditor = ({ form, onSave, newTitle }: FormFieldsEditorProps) => {
   const { data: submissions, isLoading: isSubmissionsLoading } = useFormSubmissions(form.id, 1);
   const updateForm = useUpdateForm(form.id);
-  const createForm = useCreateForm();
   const disabledFromSubmissions = (submissions ? Boolean(submissions.count) : true) && !isSubmissionsLoading;
   const disabled = updateForm.isLoading || isSubmissionsLoading || disabledFromSubmissions;
   const showSnackbar = useSnackbar();
   const [fields, setFields] = useState<Array<TextFormField | SelectFormField>>(form.fields);
   const [addButtonOpen, setAddButtonOpen] = useState(false);
   const buttonAnchorRef = useRef(null);
-  const [formtemplateName, setFormtemplateName] = useState('');
 
   useEffect(() => setFields(form.fields), [form]);
 
@@ -87,34 +69,19 @@ const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
       return;
     }
     updateForm.mutate(
-      { fields: fields, resource_type: form.resource_type },
+      { title: newTitle, fields: fields, resource_type: form.resource_type || 'Form' },
       {
         onSuccess: () => {
           showSnackbar('Spørsmålene ble oppdatert', 'success');
+          if (onSave) {
+            onSave();
+          }
         },
         onError: (e) => {
           showSnackbar(e.detail, 'error');
         },
       },
     );
-  };
-
-  const saveAsTemplate = () => {
-    const formTemplate: FormCreate = {
-      title: formtemplateName,
-      fields: removeIdsFromFields(fields),
-      resource_type: FormResourceType.FORM,
-      viewer_has_answered: false,
-      template: true,
-    };
-    createForm.mutate(formTemplate, {
-      onSuccess: (data) => {
-        showSnackbar(data.title + ' Malen ble lagret.', 'success');
-      },
-      onError: (e) => {
-        showSnackbar(e.detail, 'error');
-      },
-    });
   };
 
   return (
@@ -140,23 +107,6 @@ const FormFieldsEditor = ({ form }: FormFieldsEditorProps) => {
         <Button disabled={disabled} fullWidth onClick={save} variant='contained'>
           Lagre
         </Button>
-        <VerifyDialog
-          contentText='Hvis du lagrer som mal vil du kunne bruke malen til å opprette skjemaer senere. Gi malen en passende tittel.'
-          dialogChildren={
-            <TextField
-              disabled={disabled}
-              fullWidth
-              label='Tittel'
-              margin='normal'
-              onChange={(e) => setFormtemplateName(e.target.value)}
-              size='small'
-              value={formtemplateName}
-            />
-          }
-          onConfirm={() => saveAsTemplate()}
-          title='Lagre som mal'>
-          Lagre som mal
-        </VerifyDialog>
       </Stack>
       <Popper anchorEl={buttonAnchorRef.current} open={addButtonOpen} role={undefined} transition>
         {({ TransitionProps }) => (