+++ src/hooks/Form.tsx
@@ -10,12 +10,13 @@ import { useSnackbar } from 'hooks/Snackbar';
 export const FORM_QUERY_KEY = 'form';
 export const SUBMISSIONS_QUERY_KEY = 'submission';
 export const STATISTICS_QUERY_KEY = 'statistics';
+export const TEMPLATE_QUERY_KEY = 'templates';
 
 export const useFormById = (formId: string) =>
   useQuery<Form, RequestResponse>([FORM_QUERY_KEY, formId], () => API.getForm(formId), { enabled: formId !== '-' });
 export const useFormStatisticsById = (formId: string) =>
   useQuery<FormStatistics, RequestResponse>([FORM_QUERY_KEY, formId, STATISTICS_QUERY_KEY], () => API.getFormStatistics(formId), { enabled: formId !== '-' });
-export const useFormTemplates = () => useQuery<Array<Form>, RequestResponse>([FORM_QUERY_KEY, 'templates'], () => API.getFormTemplates(), {});
+export const useFormTemplates = () => useQuery<Array<Form>, RequestResponse>([FORM_QUERY_KEY, TEMPLATE_QUERY_KEY], () => API.getFormTemplates(), {});
 
 export const useCreateForm = (): UseMutationResult<Form, RequestResponse, FormCreate, unknown> => {
   const queryClient = useQueryClient();
@@ -45,6 +46,7 @@ export const useUpdateForm = (formId: string): UseMutationResult<Form, RequestRe
       if (data.resource_type === FormResourceType.GROUP_FORM) {
         queryClient.invalidateQueries(GROUPS_QUERY_KEYS.forms.all(data.group.slug));
       }
+      queryClient.invalidateQueries([FORM_QUERY_KEY, TEMPLATE_QUERY_KEY]);
       queryClient.invalidateQueries([FORM_QUERY_KEY, formId]);
       queryClient.setQueryData([FORM_QUERY_KEY, formId], data);
     },
@@ -58,15 +60,13 @@ export const useDeleteForm = (formId: string): UseMutationResult<RequestResponse
     onSuccess: (response) => {
       showSnackbar(response.detail, 'success');
       const data = queryClient.getQueryData<Form>([FORM_QUERY_KEY, formId]);
-      if (data?.resource_type === FormResourceType.FORM) {
-        queryClient.invalidateQueries([FORM_QUERY_KEY]);
-      }
       if (data?.resource_type === FormResourceType.EVENT_FORM) {
         queryClient.invalidateQueries([EVENT_QUERY_KEY, data.event.id]);
       }
       if (data?.resource_type === FormResourceType.GROUP_FORM) {
         queryClient.invalidateQueries(GROUPS_QUERY_KEYS.forms.all(data.group.slug));
       }
+      queryClient.invalidateQueries([FORM_QUERY_KEY, TEMPLATE_QUERY_KEY]);
       queryClient.removeQueries([FORM_QUERY_KEY, formId]);
     },
     onError: (e) => {