+++ src/pages/EventAdministration/components/EventFormAdmin.tsx
@@ -1,9 +1,9 @@
 import { EventFormCreate, Form, FormCreate, SelectFormField, SelectFormFieldOption, TextFormField } from 'types';
-import { EventFormType, FormResourceType } from 'types/Enums';
+import { EventFormType, FormFieldType, FormResourceType } from 'types/Enums';
 import { useEventById } from 'hooks/Event';
 import { useForm } from 'react-hook-form';
 import { useCreateForm, useFormTemplates, useDeleteForm } from 'hooks/Form';
-import { Typography, LinearProgress, Stack, Button } from '@mui/material';
+import { Typography, LinearProgress, Stack, Button, TextField } from '@mui/material';
 
 // Project
 import { useSnackbar } from 'hooks/Snackbar';
@@ -12,7 +12,8 @@ import FormView from 'components/forms/FormView';
 import Expand from 'components/layout/Expand';
 import FormAdmin from 'components/forms/FormAdmin';
 import { ShowMoreText } from 'components/miscellaneous/UserInformation';
-
+import FormFieldsEditor from 'components/forms/FormFieldsEditor';
+import { useState } from 'react';
 export type EventFormAdminProps = {
   eventId: number;
 };
@@ -26,20 +27,20 @@ export type EventFormEditorProps = {
 type FormTemplatePreviewType = Omit<EventFormEditorProps, 'formId'> & {
   formtemplate: Form;
 };
-
 const removeIdsFromForm = (form: Form): FormCreate => {
-  const { id, ...restForm } = form;
+  const { id, ...restForm } = form; // eslint-disable-line
   const newFields: Array<TextFormField | SelectFormField> = [];
 
   form.fields.forEach((field) => {
-    const { id, ...restField } = field;
+    const { id, ...restField } = field; // eslint-disable-line
     const newOptions: Array<SelectFormFieldOption> = [];
 
-    field.options.forEach((option) => {
-      const { id, ...restOption } = option;
-      newOptions.push(restOption as SelectFormFieldOption);
-    });
-
+    if (field.type !== FormFieldType.TEXT_ANSWER) {
+      field.options.forEach((option) => {
+        const { id, ...restOption } = option; // eslint-disable-line
+        newOptions.push(restOption as SelectFormFieldOption);
+      });
+    }
     newFields.push({ ...restField, options: newOptions } as TextFormField | SelectFormField);
   });
 
@@ -49,6 +50,8 @@ const removeIdsFromForm = (form: Form): FormCreate => {
 const FormTemplatePreview = ({ formtemplate, eventId, formType }: FormTemplatePreviewType) => {
   const deleteForm = useDeleteForm(formtemplate.id || '-');
   const { register, formState, getValues, control } = useForm<Form['fields']>();
+  const [isEditing, setIsEditing] = useState(false);
+  const [formtemplateTitle, setFormtemplateTitle] = useState(formtemplate.title);
   const createForm = useCreateForm();
   const showSnackbar = useSnackbar();
 
@@ -84,19 +87,31 @@ const FormTemplatePreview = ({ formtemplate, eventId, formType }: FormTemplatePr
   };
 
   return (
-    <>
-      <Expand flat header={formtemplate.title} sx={{ mt: 1 }}>
-        <FormView control={control} disabled={true} form={formtemplate} formState={formState} getValues={getValues} register={register} />
-        <Stack direction={{ xs: 'column', md: 'row' }} spacing={1} sx={{ p: 1 }}>
-          <Button fullWidth onClick={onCreateFormFromTemplate} variant='outlined'>
-            Bruk denne malen
-          </Button>
-          <VerifyDialog color='warning' contentText='Å slette en mal kan ikke reverseres.' onConfirm={onDeleteForm} titleText='Er du sikker?'>
-            Slett malen
-          </VerifyDialog>
-        </Stack>
-      </Expand>
-    </>
+    <Expand flat header={formtemplate.title} sx={{ mt: 1 }}>
+      {isEditing ? (
+        <>
+          <TextField label='Navn' onChange={(e) => setFormtemplateTitle(e.target.value)} type='text' value={formtemplateTitle} />
+          <FormFieldsEditor form={formtemplate} newTitle={formtemplateTitle} onSave={() => setIsEditing(false)} />
+        </>
+      ) : (
+        <>
+          <FormView control={control} disabled={true} form={formtemplate} formState={formState} getValues={getValues} register={register} />
+          <Stack>
+            <Stack direction={{ xs: 'column', md: 'row' }} spacing={1} sx={{ pb: 1 }}>
+              <Button color='primary' fullWidth onClick={() => setIsEditing(true)} variant='outlined'>
+                Rediger malen
+              </Button>
+              <VerifyDialog color='warning' contentText='Å slette en mal kan ikke reverseres.' onConfirm={onDeleteForm} titleText='Er du sikker?'>
+                Slett malen
+              </VerifyDialog>
+            </Stack>
+            <Button fullWidth onClick={onCreateFormFromTemplate} variant='outlined'>
+              Bruk denne malen
+            </Button>
+          </Stack>
+        </>
+      )}
+    </Expand>
   );
 };
 
@@ -114,7 +129,8 @@ const EventFormAdmin = ({ eventId }: EventFormAdminProps) => {
   const EventFormEditor = ({ formType }: EventFormEditorProps) => {
     const createForm = useCreateForm();
     const { data: formtemplates = [] } = useFormTemplates();
-    const newForm: EventFormCreate = {
+
+    const newForm: FormCreate = {
       title: `${event.title} - ${formType === EventFormType.SURVEY ? 'påmeldingsskjema' : 'evalueringsskjema'}`,
       type: formType,
       event: event.id,
@@ -130,12 +146,10 @@ const EventFormAdmin = ({ eventId }: EventFormAdminProps) => {
           Opprett {formType === EventFormType.SURVEY ? 'påmeldingsskjema' : 'evalueringsskjema'}
         </Button>
         <Expand flat header='Bruk en mal' sx={{ mt: 1 }}>
-          <>
-            <Typography variant='body2'>Bruk en ferdiglagd mal som utgangspunkt når du oppretter ett skjema.</Typography>
-            {formtemplates.map((formtemplate) => (
-              <FormTemplatePreview eventId={eventId} formtemplate={formtemplate} formType={formType} key={formtemplate.id} />
-            ))}
-          </>
+          <Typography variant='body2'>Bruk en ferdiglagd mal som utgangspunkt når du oppretter ett skjema.</Typography>
+          {formtemplates.map((formtemplate) => (
+            <FormTemplatePreview eventId={eventId} formtemplate={formtemplate} formType={formType} key={formtemplate.id} />
+          ))}
         </Expand>
       </>
     );